add metalosses, change base callback default to Pass
